requirements.txt
git+https://github.com/modelcontextprotocol/python-sdk.git
requests>=2.31.0
python-dotenv>=1.0.0
fastapi>=0.104.0
uvicorn>=0.24.0 